requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
brotli==1.2.0
cachetools==7.0.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.25.0
frozenlist==1.8.0
fsspec==2026.2.0
gradio==6.8.0
gradio_client==2.2.0
greenlet==3.3.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-text-splitters==1.1.1
langchainhub==0.1.21
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.8.0
pypdf==6.7.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
tomlkit==0.13.3
tqdm==4.67.3
typer==0.24.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
yarl==1.23.0
zhipuai==2.1.5.20250825
zstandard==0.25.0
